requirements.txt
torch>=1.12
torchvision
numpy
Pillow
faiss-cpu
tqdm
opencv-python
matplotlib
wandb
nbformat
